GMT_DOS-CLIENTS_FEM Fix unused warnings in fsm and asm modules
