+++ clients/fem/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "gmt_dos-clients_fem"
-version = "5.1.2"
+version = "5.1.3"
 authors.workspace = true
 edition.workspace = true
 license.workspace = true
@@ -82,6 +82,7 @@ unexpected_cfgs = { level = "warn", check-cfg = [
     'cfg(cfd2025)',
     'cfg(mount)',
     'cfg(m1)',
+    'cfg(m1_hp_force_extension)',
     'cfg(m2)',
     'cfg(topend,values("FSM","ASM"))',
     'cfg(m2_rbm,values("MCM2Lcl","MCM2Lcl6D"))',
